Fix: Unable to play recording if storage path does not exist

+++ custom_components/tapo_control/utils.py
@@ -73,11 +73,12 @@ def isOpen(ip, port):
 
 def deleteFilesOlderThan(dirPath, deleteOlderThan):
     now = datetime.datetime.utcnow().timestamp()
-    for f in os.listdir(dirPath):
-        filePath = os.path.join(dirPath, f)
-        last_modified = os.stat(filePath).st_mtime
-        if now - last_modified > deleteOlderThan:
-            os.remove(filePath)
+    if os.path.exists(dirPath):
+        for f in os.listdir(dirPath):
+            filePath = os.path.join(dirPath, f)
+            last_modified = os.stat(filePath).st_mtime
+            if now - last_modified > deleteOlderThan:
+                os.remove(filePath)
 
 
 async def getRecording(
